refactor(client): _component => _components
