+++ client/src/app/result/_components/copy-button/copy-button.tsx
