+++ client/src/app/result/_components/copy-button/index.ts
