+++ client/src/app/result/_components/index.ts
