+++ client/src/app/result/_components/misskey-share-button/index.ts
