+++ client/src/app/result/_components/misskey-share-button/misskey-share-button.css.ts
