+++ client/src/app/result/_components/modal/index.ts
