+++ client/src/app/result/_components/modal/modal.css.ts
