+++ client/src/app/result/_components/parents/index.ts
