+++ client/src/app/result/_components/parents/parents.css.ts
