+++ client/src/app/result/_components/result-image/index.ts
