+++ client/src/app/result/_components/result-image/result-image.tsx
