+++ client/src/app/result/_components/save-link/save-link.css.ts
